requirements_dev.txt
-r requirements.txt

flake8==6.0.0
Flake8-pyproject==1.2.3
factory_boy==3.2.1
mypy==1.3.0
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
tox==4.5.1
tox-gh-actions==3.1.1
